feature/router-integration
- add routes for toggle
