+++ src/app/AppRoutes.tsx
@@ -3,14 +3,16 @@ import App from './App.tsx';
 import About from '../entities/App/ui/About.tsx';
 import NotFound from '../entities/App/ui/NotFound.tsx';
 import Layout from '../entities/App/ui/Layout.tsx';
+import Auth from '../entities/User/ui/Auth.tsx';
 
 const AppRoutes = () => {
 	return (
 		<Routes>
 			<Route element={<Layout />}>
-				<Route index element={<App />} />
+				<Route path={'/'} index element={<App />} />
 				<Route path={'/about'} element={<About />} />
 				<Route path={'*'} element={<NotFound />} />
+				<Route path={'/auth/*'} element={<Auth />} />
 			</Route>
 		</Routes>
 	);
